N21-931 renamed and moved external-tool-parameter-validation.service.ts to external tool
