+++ apps/server/src/modules/tool/common/service/index.ts
@@ -1,2 +1,1 @@
 export * from './common-tool.service';
-export * from './common-tool-validation.service';
