+++ apps/server/src/modules/tool/external-tool/service/index.ts
@@ -2,3 +2,4 @@ export * from './external-tool.service';
 export * from './external-tool-service.mapper';
 export * from './external-tool-version.service';
 export * from './external-tool-validation.service';
+export * from './external-tool-parameter-validation.service';
